Refactor to prepare for unit testing and clean up eventLoader code.
